Added cli requirements to circleci key hashes; restructured tests within /contrib
