+++ .circleci/config.yml
@@ -9,7 +9,7 @@ jobs:
       - checkout
       - restore_cache:
           keys:
-            - v1-dependencies-20.0-{{ checksum "requirements.txt" }}-{{ checksum "requirements-dev.txt" }}-{{ checksum "contrib/client/requirements.txt" }}}
+            - v1-dependencies-20.0-{{ checksum "requirements.txt" }}-{{ checksum "requirements-dev.txt" }}-{{ checksum "teos/cli/requirements.txt" }}-{{ checksum "contrib/client/requirements.txt" }}}
 
       - run:
           name: Install dependencies
@@ -35,7 +35,7 @@ jobs:
           paths:
             - ./venv
             - bitcoin
-          key: v1-dependencies-20.0-{{ checksum "requirements.txt" }}-{{ checksum "requirements-dev.txt" }}-{{ checksum "contrib/client/requirements.txt" }}}
+          key: v1-dependencies-20.0-{{ checksum "requirements.txt" }}-{{ checksum "requirements-dev.txt" }}-{{ checksum "teos/cli/requirements.txt" }}-{{ checksum "contrib/client/requirements.txt" }}}
 
       - run:
           name: Setup teos
@@ -47,7 +47,7 @@ jobs:
           name: Run contrib unit tests
           command: |
             . venv/bin/activate
-            pytest contrib/test
+            pytest contrib/**/test
 
       - run:
           name: Run common unit tests
